+++ com.unity.render-pipelines.high-definition/Editor/VFXGraph/Outputs/VFXDistortionQuadOutput.cs
@@ -2,26 +2,39 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEditor.VFX.Block;
 using UnityEngine;
-using UnityEngine.Experimental.VFX;
 
 namespace UnityEditor.VFX
 {
     [VFXInfo]
-    class VFXLitMeshOutput : VFXAbstractParticleHDRPLitOutput
+    class VFXDistortionQuadOutput : VFXAbstractDistortionOutput
     {
-        public override string name { get { return "Lit Mesh Output"; } }
-        public override string codeGeneratorTemplate { get { return RenderPipeTemplate("VFXParticleLitMesh"); } }
-        public override VFXTaskType taskType { get { return VFXTaskType.ParticleMeshOutput; } }
+        //[VFXSetting] // tmp dont expose as settings atm
+        public bool useGeometryShader = false;
+
+        public override string name { get { return "Distortion Quad Output"; } }
+        public override string codeGeneratorTemplate { get { return RenderPipeTemplate("VFXParticleDistortionQuad"); } }
+        public override VFXTaskType taskType { get { return useGeometryShader ? VFXTaskType.ParticlePointOutput : VFXTaskType.ParticleQuadOutput; } }
         public override bool supportsUV { get { return true; } }
-        public override CullMode defaultCullMode { get { return CullMode.Back; } }
+
+        public override IEnumerable<string> additionalDefines
+        {
+            get
+            {
+                foreach (var def in base.additionalDefines)
+                    yield return def;
+
+                if (useGeometryShader)
+                    yield return "USE_GEOMETRY_SHADER";
+
+                yield return "VFX_PRIMITIVE_QUAD";
+            }
+        }
 
         public override IEnumerable<VFXAttributeInfo> attributes
         {
             get
             {
                 yield return new VFXAttributeInfo(VFXAttribute.Position, VFXAttributeMode.Read);
-                if (colorMode != ColorMode.None)
-                    yield return new VFXAttributeInfo(VFXAttribute.Color, VFXAttributeMode.Read);
                 yield return new VFXAttributeInfo(VFXAttribute.Alpha, VFXAttributeMode.Read);
                 yield return new VFXAttributeInfo(VFXAttribute.Alive, VFXAttributeMode.Read);
                 yield return new VFXAttributeInfo(VFXAttribute.AxisX, VFXAttributeMode.Read);
@@ -42,35 +55,6 @@ namespace UnityEditor.VFX
                 if (usesFlipbook)
                     yield return new VFXAttributeInfo(VFXAttribute.TexIndex, VFXAttributeMode.Read);
             }
-        }
-
-        public class InputProperties
-        {
-            [Tooltip("Mesh to be used for particle rendering.")]
-            public Mesh mesh = VFXResources.defaultResources.mesh;
-            [Tooltip("Define a bitmask to control which submeshes are rendered.")]
-            public uint subMeshMask = 0xffffffff;
-        }
-
-        public override VFXExpressionMapper GetExpressionMapper(VFXDeviceTarget target)
-        {
-            var mapper = base.GetExpressionMapper(target);
-
-            switch (target)
-            {
-                case VFXDeviceTarget.CPU:
-                {
-                    mapper.AddExpression(inputSlots.First(s => s.name == "mesh").GetExpression(), "mesh", -1);
-                    mapper.AddExpression(inputSlots.First(s => s.name == "subMeshMask").GetExpression(), "subMeshMask", -1);
-                    break;
-                }
-                default:
-                {
-                    break;
-                }
-            }
-
-            return mapper;
         }
     }
 }