+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -9,13 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [6.8.0-preview] - 2019-06-12
 
 ### Added
-- Subgraphs
 - Make multiselection work in a way that do not assume that the same parameter will have the same index in the property sheet.
 - auto recompile when changing shaderpath
 - auto recompile new vfx
 - better detection of default shader path
-- Bitfield control
-- Initial Event Name inspector for visual effect asset and component
+- Bitfield control.
+- Move HDRP outputs to HDRP package + expose HDRP queue selection
+- Add exposure weight control for HDRP outputs
 - Subgraphs
 
 ## [6.7.0-preview] - 2019-05-21